+++ custom.js
@@ -43,6 +43,7 @@ var THEME_MODULES = [
     "src/js/scheme.js",
     "src/js/schemes.js",
     "src/js/color-oklch.js",
+    "src/js/color-repair.js",
     "src/js/scheme-generator.js",
     "src/js/iconpack.js",
     "src/js/search.js",